props/specimens/ducktape_llm_common/2026-01-03-00/code/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ducktape-llm-common"
version = "0.1.0"
description = "Common utilities, linters, and prompts for LLM development workflows"
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
]
keywords = ["linter", "pre-commit", "workflow", "llm", "ai", "automation"]
dependencies = [
    "PyYAML>=6.0",
    "click>=8.0",
    "humanize>=4.0.0",  # for human-readable time formatting
    "platformdirs>=3.0",
    "pydantic>=2.0",
    "pygit2>=1.18.0",
    "pytimeparse>=1.1.8",  # for parsing duration strings like "2h", "30m"
    "ruff>=0.1.0",  # actual prod dependency of linter
    "pre-commit>=3.0",  # for claude-linter CLI hook
    "tomli>=2.0.0",  # for reading TOML files
    "tomli-w>=1.0.0",  # for writing TOML files
]

[project.optional-dependencies]
notifications = [
    "dbus-python>=1.2.0",  # for desktop notifications (Linux only)
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "PyHamcrest>=2.1",
]

[project.scripts]
claude-linter = "ducktape_llm_common.claude_linter.cli:cli"
claude-linter-v2 = "ducktape_llm_common.claude_linter_v2.cli:cli"
cl2 = "ducktape_llm_common.claude_linter_v2.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["ducktape_llm_common"]

[tool.hatch.build]
include = [
    "ducktape_llm_common/**/*.py",
    "ducktape_llm_common/**/*.sh",
]

[tool.hatch.build.targets.sdist]
include = [
    "/ducktape_llm_common",
    "/README.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=ducktape_llm_common --cov-report=term-missing"
norecursedirs = ["__pycache__", "*.egg", "dist", "build", ".git"]


[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
